refactor(fig): own Figma clipboard encoding and conversion
