+++ packages/fig/package.json
@@ -10,6 +10,12 @@
       "import": "./dist/index.js",
       "default": "./dist/index.js"
     },
+    "./clipboard": {
+      "types": "./dist/clipboard.d.ts",
+      "bun": "./src/clipboard/index.ts",
+      "import": "./dist/clipboard.js",
+      "default": "./dist/clipboard.js"
+    },
     "./node-change": {
       "types": "./dist/node-change.d.ts",
       "bun": "./src/node-change/index.ts",
@@ -54,7 +60,8 @@
   },
   "dependencies": {
     "es-toolkit": "^1.51.0",
-    "fflate": "^0.8.3"
+    "fflate": "^0.8.3",
+    "js-base64": "^3.9.3"
   },
   "devDependencies": {
     "@types/bun": "^1.3.14",
