+++ packages/fig/tsdown.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   entry: {
     index: './src/index.ts',
     'instance-overrides': './src/instance-overrides/index.ts',
+    clipboard: './src/clipboard/index.ts',
     'node-change': './src/node-change/index.ts'
   },
   platform: 'neutral',
